Issue:
It seems that this change would solve an issue with the callback#finishedLoading
method invocation when the ressource was already loaded.
But still, the callback is never being called.
It's like the resource is never loaded.

How to reproduce the issue:

Go to mockup view/click New project ~> while in the new project view the
Model's project hasn't been initialized nor created.