Invoke own version of jsonschema2pojo, using LibGDX Array
Use es.e-ucm:jsonschema2pojo-maven-plugin, which allows
arrayDefinition and arrayImplementation classes to be specified

Updated all accesses to the schema to refer to Arrays instead
of List or ArrayList. There is mostly a 1-to-1 translation.

Also changed the engine's schema pom, because LibGDX now has to be
available in the schema.